Shulker Box not dropping fix
