+++ src/main/resources/mixins.tweakeroo.json
@@ -1,82 +1,83 @@
 {
-        "required": true,
-        "package": "fi.dy.masa.tweakeroo.mixin",
-        "minVersion": "0.8",
-        "compatibilityLevel": "JAVA_17",
-        "client": [
-                "IMixinAbstractBlock",
-                "IMixinAbstractHorseEntity",
-                "IMixinAxeItem",
-                "IMixinChunkLightProvider",
-                "IMixinClientWorld",
-                "IMixinCommandBlockExecutor",
-                "IMixinShovelItem",
-                "IMixinSimpleOption",
-                "MixinAbstractClientPlayerEntity",
-                "MixinAbstractInventoryScreen",
-                "MixinAbstractSignEditScreen",
-                "MixinBackgroundRenderer",
-                "MixinBatEntity",
-                "MixinBeaconBlockEntityRenderer",
-                "MixinBlockEntityRenderDispatcher",
-                "MixinBlockItem",
-                "MixinBossBarHud",
-                "MixinBuiltChunk",
-                "MixinChatHud",
-                "MixinChatScreen",
-                "MixinChunkBuilder_BuiltChunk",
-                "MixinClientBossBar",
-                "MixinClientCommandSource",
-                "MixinClientPlayerEntity",
-                "MixinClientPlayerInteractionManager",
-                "MixinClientPlayNetworkHandler",
-                "MixinClientWorld",
-                "MixinClientWorld_Properties",
-                "MixinCloneCommand",
-                "MixinCommandBlockScreen",
-                "MixinCreativeInventoryScreen",
-                "MixinDimensionEffects_Nether",
-                "MixinEntity",
-                "MixinEntityRenderDispatcher",
-                "MixinExplosion",
-                "MixinFillCommand",
-                "MixinGameRenderer",
-                "MixinHeldItemRenderer",
-                "MixinInGameHud",
-                "MixinItem",
-                "MixinItemEntity",
-                "MixinItemGroup",
-                "MixinItemStack",
-                "MixinKeyboardInput",
-                "MixinLightingProvider",
-                "MixinLivingEntity",
-                "MixinMinecraftClient",
-                "MixinMobSpawnerBlockEntityRenderer",
-                "MixinMobSpawnerLogic",
-                "MixinMouse",
-                "MixinNetherPortalBlock",
-                "MixinObserverBlock",
-                "MixinParticleManager",
-                "MixinPlayerAbilities",
-                "MixinPlayerEntity",
-                "MixinRavagerEntity",
-                "MixinScaffoldingBlock",
-                "MixinSculkSensor",
-                "MixinServerPlayNetworkHandler",
-                "MixinShulkerBoxBlock",
-                "MixinSignBlockEntity",
-                "MixinSlimeBlock",
-                "MixinStructureBlockBlockEntity",
-                "MixinTeleportSpectatorMenu",
-                "MixinThreadedAnvilChunkStorage",
-                "MixinTradeOffer",
-                "MixinUpdateStructureBlockC2SPacket",
-                "MixinWindow",
-                "MixinWorld",
-                "MixinWorldRenderer"
-        ],
-        "mixinPriority": 990,
-        "injectors": {
-                "defaultRequire": 0
+    "required": true,
+    "package": "fi.dy.masa.tweakeroo.mixin",
+    "minVersion": "0.8",
+    "compatibilityLevel": "JAVA_17",
+    "client": [
+        "IMixinAbstractBlock",
+        "IMixinAbstractHorseEntity",
+        "IMixinAxeItem",
+        "IMixinChunkLightProvider",
+        "IMixinClientWorld",
+        "IMixinCommandBlockExecutor",
+        "IMixinShovelItem",
+        "IMixinSimpleOption",
+        "MixinAbstractClientPlayerEntity",
+        "MixinAbstractInventoryScreen",
+        "MixinAbstractSignEditScreen",
+        "MixinBackgroundRenderer",
+        "MixinBatEntity",
+        "MixinBeaconBlockEntityRenderer",
+        "MixinBlockEntityRenderDispatcher",
+        "MixinBlockItem",
+        "MixinBossBarHud",
+        "MixinBuiltChunk",
+        "MixinChatHud",
+        "MixinChatScreen",
+        "MixinChunkBuilder_BuiltChunk",
+        "MixinClientBossBar",
+        "MixinClientCommandSource",
+        "MixinClientPlayerEntity",
+        "MixinClientPlayerInteractionManager",
+        "MixinClientPlayNetworkHandler",
+        "MixinClientWorld",
+        "MixinClientWorld_Properties",
+        "MixinCloneCommand",
+        "MixinCommandBlockScreen",
+        "MixinCreativeInventoryScreen",
+        "MixinDimensionEffects_Nether",
+        "MixinEntity",
+        "MixinEntityRenderDispatcher",
+        "MixinExplosion",
+        "MixinFillCommand",
+        "MixinGameRenderer",
+        "MixinGameRenderer_ViewBob",
+        "MixinHeldItemRenderer",
+        "MixinInGameHud",
+        "MixinItem",
+        "MixinItemEntity",
+        "MixinItemGroup",
+        "MixinItemStack",
+        "MixinKeyboardInput",
+        "MixinLightingProvider",
+        "MixinLivingEntity",
+        "MixinMinecraftClient",
+        "MixinMobSpawnerBlockEntityRenderer",
+        "MixinMobSpawnerLogic",
+        "MixinMouse",
+        "MixinNetherPortalBlock",
+        "MixinObserverBlock",
+        "MixinParticleManager",
+        "MixinPlayerAbilities",
+        "MixinPlayerEntity",
+        "MixinRavagerEntity",
+        "MixinScaffoldingBlock",
+        "MixinSculkSensor",
+        "MixinServerPlayNetworkHandler",
+        "MixinShulkerBoxBlock",
+        "MixinSignBlockEntity",
+        "MixinSlimeBlock",
+        "MixinStructureBlockBlockEntity",
+        "MixinTeleportSpectatorMenu",
+        "MixinThreadedAnvilChunkStorage",
+        "MixinTradeOffer",
+        "MixinUpdateStructureBlockC2SPacket",
+        "MixinWindow",
+        "MixinWorld",
+        "MixinWorldRenderer"
+    ],
+    "mixinPriority": 990,
+    "injectors": {
+        "defaultRequire": 0
         }
 }
